create image-tracker
